refactor [James]: update log interface and model
