+++ server-side/src/controllers/admin.controller.js
@@ -4,7 +4,7 @@ const MembershipHistory = require("../models/history,model");
 const Merch = require("../models/merch.model");
 const Student = require("../models/StudentModel");
 const Order = require("../models/OrdersModel");
-const Log = require("../models/LogModel");
+const Log = require("../models/log.model");
 const { format, startOfDay, endOfDay } = require("date-fns");
 const { admin_model, role_model } = require("../model_template/model_data");
 const { membershipRequestReceipt } = require("../mail_template/mail.template");
