+++ server-side/src/controllers/merchandise.controller.js
@@ -2,7 +2,7 @@ const Merch = require("../models/merch.model");
 const Student = require("../models/StudentModel");
 const Orders = require("../models/OrdersModel");
 const Admin = require("../models/AdminModel");
-const Log = require("../models/LogModel");
+const Log = require("../models/log.model");
 const Event = require("../models/EventsModel");
 const mongoose = require("mongoose");
 const { DeleteObjectCommand } = require("@aws-sdk/client-s3");
